// src/app/models/match.ts
export class Match {
    competition_id?: Number;
    match_id?: Number;
    event_id?: Number;
    event_type_id?: Number;
    market_id?: Number;
    cricbuzz_id?: String;
    long_name?: String;
    short_name?: String;
    start_time?: String;
    bet_min?: Number;
    bet_max?: Number;
    teams?: any;
}